Completed roll 8 dice feature
I completed the roll a dice feature as well as the rolling 8 die feature and had it print the results onto the console. I logged this into my logbook and added it onto my backlog.

For the code, I commented out the debug print statements for the dice class and isolated the section needed to roll the 8 die in the main class and stored it into an array for easy access.